+++ src/ComponentPresenter/index.js
@@ -1,0 +1,3 @@
+import ComponentPresenter from './ComponentPresenter';
+
+export default ComponentPresenter;