+++ src/components/Button/Showcase.js
@@ -6,19 +6,19 @@ import Button from './Button';
 const ButtonShowcase = () => (
   <div>
     <Button text="default variant" />
-    <br />
+    <br /><br />
     <Button text="default variant small" small />
-    <br />
+    <br /><br />
     <Button text="default variant light" light bold={false} />
-    <br />
+    <br /><br />
     <Button text="default variant !bold" bold={false} />
-    <br />
+    <br /><br />
     <Button text="filled variant" filled bold={false} />
-    <br />
+    <br /><br />
     <Button text="filled variant bold" filled />
-    <br />
+    <br /><br />
     <Button text="filled variant inverted" filled inverted bold={false} />
-    <br />
+    <br /><br />
     <Button text="filled variant inverted bold" filled inverted />
   </div>
 );