+++ src/components/Slider/index.js
@@ -1,0 +1,3 @@
+import Slider from 'components/Slider/Slider';
+
+export default Slider;