+++ CHANGELOG.md
@@ -1,36 +1,73 @@
 # Changelog
+
 > generated with [@semantic-release/changelog](https://github.com/semantic-release/changelog)
+
+# [1.4.0](https://github.com/viewar/components/compare/v1.3.0...v1.4.0) (2019-11-20)
+
+### Bug Fixes
+
+- **ButtonToggle:** allow `<ButtonToggle />` to be un-controlled ([bc552bc](https://github.com/viewar/components/commit/bc552bc4ad6d566f37dafd09456c4945c66fdcca))
+- **scripts:** fix 'format' - ignore build ([c6d6919](https://github.com/viewar/components/commit/c6d691916c445f562bf40e700a926fee98a6bb98))
+
+### Features
+
+- **ButtonToggle:** add `<ButtonToggle />` ([8c7f1b3](https://github.com/viewar/components/commit/8c7f1b3f84175d1e0e15199ca1fe13d57b575ddf))
+- **Slider:** add `<Slider />` component + Showcase ([#19](https://github.com/viewar/components/issues/19)) ([e22b30e](https://github.com/viewar/components/commit/e22b30e4e5f264f46ced7bbafa164457079a1037))
+- **PropertyPicker:** only show 'item.properties' filtered by 'props.shownProperties' ([891c5b5](https://github.com/viewar/components/commit/891c5b54641ba6f27f4e8642824b915f1c5659d0))
+- **componentPresenter:** render Showcases per router param ([#20](https://github.com/viewar/components/issues/20)) ([43730b5](https://github.com/viewar/components/commit/43730b56446f59d7cedf04de0f98cecb390b8ac6))
+
+# [1.3.0](https://github.com/viewar/components/compare/v1.2.1...v1.3.0) (2019-11-18)
+
+### Features
+
+- **spec:** add minimal spec test for `<Overlay />` ([b2bd667](https://github.com/viewar/components/commit/b2bd667896398213def303d5ca8b5536cedc2048))
+- **spec:** add spec test for `<Input />` ([795a643](https://github.com/viewar/components/commit/795a6434a4b6a53eeec121c393ab0f0e4e1f45d4))
+- **spec:** add spec test for `<SideBar />` ([b0fa38e](https://github.com/viewar/components/commit/b0fa38e125c345220a99d1cea2bb3103c0af811e))
+
+## [1.2.1](https://github.com/viewar/components/compare/v1.2.0...v1.2.1) (2019-11-17)
+
+### Bug Fixes
+
+- **CI**: use "Build Stages" for build and deployment (#14) ([3ae0574](https://github.com/viewar/components/commit/3ae0574e708f8539738d775f9d8d24e3fe014866))
+- **CHANGELOG:** add initial components to log and fix duplicate entries ([5317612](https://github.com/viewar/components/commit/5317612be72918611ae8831a0797e7a5efe652e5))
 
 # [1.2.0](https://github.com/viewar/components/compare/v1.1.0...v1.2.0) (2019-11-16)
 
-
 ### Bug Fixes
 
-* **Accordion:** fix propType for children ([175e38f](https://github.com/viewar/components/commit/175e38f3f6f192f0e622b1efd36c3220b84eec2f))
-* **Accordion:** fix propType for children ([0682674](https://github.com/viewar/components/commit/06826744b4469c22ef7312ac09ce2be762f48681))
-
+- **Accordion:** fix propType for children ([175e38f](https://github.com/viewar/components/commit/175e38f3f6f192f0e622b1efd36c3220b84eec2f))
 
 ### Features
 
-* **spec:** add 'App.spec.js' ([4a7bc92](https://github.com/viewar/components/commit/4a7bc92c7fa4ed7bda30df27eefcc77dfe082e56))
-* **spec:** add 'App.spec.js' ([7037976](https://github.com/viewar/components/commit/703797652fe667cb2ddc68033030132c06d059de))
-* **viewar:** add 'viewar-core' and 'viewar-api' ([#12](https://github.com/viewar/components/issues/12)) ([25eb40a](https://github.com/viewar/components/commit/25eb40a7c4dc62e5a95048107ae7cb53774804e4))
-* **viewar:** add 'viewar-core' and 'viewar-api' ([#12](https://github.com/viewar/components/issues/12)) ([a964d86](https://github.com/viewar/components/commit/a964d8622086ab15702d5909ddfe4eb7dc54f12a))
-* **spec:** add spec test for `<Accordion />` ([d86962a](https://github.com/viewar/components/commit/d86962ac0b6b4b310fe8cd6630793dd1b4f2e963))
-* **spec:** add spec test for `<Accordion />` ([fa7edc2](https://github.com/viewar/components/commit/fa7edc279840698d24aa55f7254f3ec6850c49e1))
+- **viewar:** add 'viewar-core' and 'viewar-api' ([#12](https://github.com/viewar/components/issues/12)) ([a964d86](https://github.com/viewar/components/commit/a964d8622086ab15702d5909ddfe4eb7dc54f12a))
 
-# Changelog
+### Components
 
-> generated with [@semantic-release/changelog](https://github.com/semantic-release/changelog)
+- **spec:** add 'App.spec.js' ([4a7bc92](https://github.com/viewar/components/commit/4a7bc92c7fa4ed7bda30df27eefcc77dfe082e56))
+- **spec:** add spec test for `<Accordion />` ([d86962a](https://github.com/viewar/components/commit/d86962ac0b6b4b310fe8cd6630793dd1b4f2e963))
 
 # [1.1.0](https://github.com/viewar/components/compare/v1.0.0...v1.1.0) (2019-11-15)
 
-### Features
+### Components
 
 - **PropertyPicker:** add `<PropertyPicker />` ([a85ccf9](https://github.com/viewar/components/commit/a85ccf9c1bc261c5437d545d9102598b456d7656))
+- **SceneEditor:** add `<SceneEditor />` ([1f7fc92](https://github.com/viewar/components/commit/1f7fc92d6b52f0822557b0f12be8cebaeec35005))
 
 # [1.0.0](https://github.com/viewar/components/compare/1a5ba55...v1.0.0) (2019-11-12)
 
-### Bug Fixes
+### Features
 
 - **setup:** stable setup ([v1.0.0](https://github.com/viewar/components/tree/v1.0.0))
+  - add '@babel' modules
+  - use '@viewar/webpack' and '@viewar/config-eslint'
+
+### Components
+
+- **Button:** add `<Button />` ([1f7fc92](https://github.com/viewar/components/commit/1f7fc92d6b52f0822557b0f12be8cebaeec35005))
+- **Input:** add `<Input />` ([1f7fc92](https://github.com/viewar/components/commit/1f7fc92d6b52f0822557b0f12be8cebaeec35005))
+- **Accordion:** add `<Accordion />` ([1f7fc92](https://github.com/viewar/components/commit/1f7fc92d6b52f0822557b0f12be8cebaeec35005))
+- **Button:** add `<Button />` ([1f7fc92](https://github.com/viewar/components/commit/1f7fc92d6b52f0822557b0f12be8cebaeec35005))
+- **Icon:** add `<Icon />` ([1f7fc92](https://github.com/viewar/components/commit/1f7fc92d6b52f0822557b0f12be8cebaeec35005)
+- **SideBar:** add `<SideBar />` ([1f7fc92](https://github.com/viewar/components/commit/1f7fc92d6b52f0822557b0f12be8cebaeec35005))
+- **Overlay:** add `<Overlay />` ([1f7fc92](https://github.com/viewar/components/commit/1f7fc92d6b52f0822557b0f12be8cebaeec35005))
+- **LoadingState:** add `<LoadingState />` ([1f7fc92](https://github.com/viewar/components/commit/1f7fc92d6b52f0822557b0f12be8cebaeec35005))